Webpack v4 compatibility
In my limited testing as a new webpack user, this plugin already works fine with webpack v4 so all that's needed is to squash the deprecation warning. I expect this change to be fully backwards compatible with older webpack versions.

I'm making this change using the Github web interface so please excuse any silly mistakes or lack of testing!